added masterlist
same dedupe script as previouly, however it also adds a master list for deduplicated records into the results folder. the master list will be created from scratch if it does not exist, or new data will be appended.

Important: need to change run all R file to add the database ID to the all_results_tmp file, otherwise this will not work.